First workable version of favorites downloading has been completed.
If you wish to use it, download the fav_search ini file as well. This will allow you to enter a username as the query, and set the favorites download to True, allowing you to pick from a list of a users collections and download all images within that collection.